# MWPW-141594 Move georouting to post LCP