@auniverseaway 2024-01-30
Thanks for taking the time to check this out and get it to the finish line.

Probably worth mentioning to the martech folks when this gets merged as it will slightly delay martech compared to what we do today. This is because we are now awaiting GeoRouting and putting it in front of martech.