@meganthecoder 2024-01-31
> Marking as "request-changes" to block merge as I need to do some additional perf testing.

@chrischrischris Sounds good. Let me know if I can help at all. This is a high priority for us